% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sampleprob}
\alias{sampleprob}
\title{Sample Prob}
\usage{
sampleprob(url = "http://dart.fss.or.kr/dsaf001/main.do?rcpNo=20200330003851")
}
\arguments{
\item{url}{sample URL from DART system}
}
\description{
example of function containing Korean
}
